Splitting up non-geometry board search library in preparation for adding a new component search
